+++ app/context/index.js
@@ -94,11 +94,16 @@ class Context extends EE {
 
     get contextArray() { return this.#contextArray; }
     get featureArray() { return this.#featureArray; }
+    get features() { return this.#featureArray; }
     get filterArray() { return this.#filterArray; }
+    get filters() { return this.#filterArray; }
+
+    // List all roles linked to this context
+    get roles() {}
 
     // List all apps linked to this context
     get apps() {}
-
+    
     // List all identities linked to this context
     get identities() {}
 
@@ -220,6 +225,24 @@ class Context extends EE {
         filArr = this.#filterArray) {
 
     }
+
+    /**
+     * Features
+     */    
+
+    setContextFeatures(idOrArray) {}
+    addFeatureToContext(idOrArray) {}
+    removeFeatureFromContext(idOrArray) {}
+    listActiveContextFeatures() {}
+
+    /**
+     * Filters
+     */    
+    
+    setContextFilters(idOrArray) {}
+    addFilterToContext(idOrArray) {}
+    removeFilterFromContext(idOrArray) {}
+    listActiveContextFilters() {}
 
 
     /**
